Move Editor to components
